+++ am-kernels/tests/cpu-tests/tests/4s-shift.c
@@ -4,19 +4,19 @@ int main(){
     signed int ans = 0;
     for(signed int a = -2147483648; a < 2147483647; a = a + 1){
         for(signed int b = -2147483648; b < 2147483647; b = b + 1){
-            ans = (unsigned int)a << (unsigned int)b;
+            ans = (unsigned int)a << b;
             //printf("%d, ", ans);
         }
     }
     for(signed int a = -2147483648; a < 2147483647; a = a + 1){
         for(signed int b = -2147483648; b < 2147483647; b = b + 1){
-            ans = (unsigned int)a >> (unsigned int)b;
+            ans = (unsigned int)a >> b;
             //printf("%d, ", ans);
         }
     }
     for(signed int a = -2147483648; a < 2147483647; a = a + 1){
         for(signed int b = -2147483648; b < 2147483647; b = b + 1){
-            ans = (signed int)a >> (signed int)b;
+            ans = (signed int)a >> b;
             //printf("%d, ", ans);
         }
     }
