+++ am-kernels/tests/cpu-tests/tests/4u-shift.c
@@ -4,19 +4,19 @@ int main(){
     unsigned int ans = 0;
     for(unsigned int a = 0; a < 4294967295; a = a + 1){
         for(unsigned int b = 0; b < 4294967295; b = b + 1){
-            ans = (unsigned int)a << (unsigned int)b;
+            ans = (unsigned int)a << b;
             //printf("%d, ", ans);
         }
     }
     for(unsigned int a = 0; a < 4294967295; a = a + 1){
         for(unsigned int b = 0; b < 4294967295; b = b + 1){
-            ans = (unsigned int)a >> (unsigned int)b;
+            ans = (unsigned int)a >> b;
             //printf("%d, ", ans);
         }
     }
     for(unsigned int a = 0; a < 4294967295; a = a + 1){
         for(unsigned int b = 0; b < 4294967295; b = b + 1){
-            ans = (signed int)a >> (signed int)b;
+            ans = (signed int)a >> b;
             //printf("%d, ", ans);
         }
     }
